backend/requirements.txt
fastapi==0.115.0
uvicorn==0.29.0
requests==2.31.0
tldextract==5.1.1
python-whois==0.9.4
python-dotenv==1.0.1
anthropic==0.25.0
aiohttp==3.9.5
pydantic==2.6.4
pydantic-core==2.16.3
